Move public HTTP server to assemble handlers in main.go
