+++ README.md
@@ -49,10 +49,4 @@ Usage of ./observatorium:
     	The address on which internal server runs. (default ":8081")
   -web.listen string
     	The address on which public server runs. (default ":8080")
-  -web.timeout duration
-    	The maximum duration before timing out the request, and closing idle connections. (default 2m0s)
-  -web.timeout.read duration
-    	The maximum duration before reading the entire request, including the body. (default 2m0s)
-  -web.timeout.write duration
-    	The maximum duration  before timing out writes of the response. (default 2m0s)
 ```
